+++ clang/test/SemaCXX/cxx1z-constexpr-lambdas.cpp
@@ -6,7 +6,7 @@
 
 namespace test_lambda_is_literal {
 #ifdef CPP14_AND_EARLIER
-//expected-error@+4{{not a literal type}}
+//expected-error@+4{{non-literal parameter type}}
 //expected-note@+2{{lambda closure types are non-literal types before C++17}}
 #endif
 auto L = [] { };
@@ -19,7 +19,7 @@ namespace test_constexpr_checking {
 
 namespace ns1 {
   struct NonLit { ~NonLit(); };  //expected-note{{not literal}}
-  auto L = [](NonLit NL) constexpr { }; //expected-error{{not a literal type}}
+  auto L = [](NonLit NL) constexpr { }; //expected-error{{non-literal parameter type}}
 } // end ns1
 
 namespace ns2 {